+++ src/Database/V5/Bloodhound/Types.hs
@@ -205,6 +205,7 @@ module Database.V5.Bloodhound.Types
        , Analyzer(..)
        , Tokenizer(..)
        , TokenFilter(..)
+       , CharFilter(..)
        , MaxExpansions(..)
        , Lenient(..)
        , MatchQueryType(..)
@@ -387,6 +388,7 @@ module Database.V5.Bloodhound.Types
        , AnalyzerDefinition(..)
        , TokenizerDefinition(..)
        , TokenFilterDefinition(..)
+       , CharFilterDefinition(..)
        , Ngram(..)
        , TokenChar(..)
        , Shingle(..)
@@ -417,6 +419,7 @@ import qualified Data.Map.Strict                       as M
 import           Data.Maybe
 import           Data.Scientific                       (Scientific)
 import           Data.Semigroup
+import           Data.String                           (IsString)
 import           Data.Text                             (Text)
 import qualified Data.Text                             as T
 import           Data.Time.Calendar
@@ -607,13 +610,15 @@ data Analysis = Analysis
   { analysisAnalyzer :: M.Map Text AnalyzerDefinition
   , analysisTokenizer :: M.Map Text TokenizerDefinition
   , analysisTokenFilter :: M.Map Text TokenFilterDefinition
+  , analysisCharFilter :: M.Map Text CharFilterDefinition
   } deriving (Eq,Show,Generic,Typeable)
 
 instance ToJSON Analysis where
-  toJSON (Analysis analyzer tokenizer tokenFilter) = object
+  toJSON (Analysis analyzer tokenizer tokenFilter charFilter) = object
     [ "analyzer" .= analyzer
     , "tokenizer" .= tokenizer
     , "filter" .= tokenFilter
+    , "char_filter" .= charFilter
     ]
 
 instance FromJSON Analysis where
@@ -621,23 +626,62 @@ instance FromJSON Analysis where
     <$> m .: "analyzer"
     <*> m .:? "tokenizer" .!= M.empty
     <*> m .:? "filter" .!= M.empty
+    <*> m .:? "char_filter" .!= M.empty
 
 data AnalyzerDefinition = AnalyzerDefinition
   { analyzerDefinitionTokenizer :: Maybe Tokenizer
   , analyzerDefinitionFilter :: [TokenFilter]
+  , analyzerDefinitionCharFilter :: [CharFilter]
   } deriving (Eq,Show,Generic,Typeable)
 
 instance ToJSON AnalyzerDefinition where
-  toJSON (AnalyzerDefinition tokenizer tokenFilter) = object $ catMaybes
+  toJSON (AnalyzerDefinition tokenizer tokenFilter charFilter) = object $ catMaybes
     [ fmap ("tokenizer" .=) tokenizer
     , Just $ "filter" .= tokenFilter
+    , Just $ "char_filter" .= charFilter
     ]
 
 instance FromJSON AnalyzerDefinition where
   parseJSON = withObject "AnalyzerDefinition" $ \m -> AnalyzerDefinition
     <$> m .:? "tokenizer"
     <*> m .:? "filter" .!= []
-  
+    <*> m .:? "char_filter" .!= []
+
+-- | Character filters are used to preprocess the stream of characters
+--   before it is passed to the tokenizer.
+data CharFilterDefinition
+  = CharFilterDefinitionMapping (M.Map Text Text)
+  | CharFilterDefinitionPatternReplace
+    { charFilterDefinitionPatternReplacePattern :: Text
+    , charFilterDefinitionPatternReplaceReplacement :: Text
+    , charFilterDefinitionPatternReplaceFlags :: Maybe Text
+    }
+  deriving (Eq,Show,Generic)
+
+instance ToJSON CharFilterDefinition where
+  toJSON (CharFilterDefinitionMapping ms) = object
+    [ "type" .= ("mapping" :: Text)
+    , "mappings" .= [a <> " => " <> b | (a, b) <- M.toList ms] ]
+  toJSON (CharFilterDefinitionPatternReplace pat repl flags) = object $
+    [ "type" .= ("pattern_replace" :: Text)
+    , "pattern" .= pat
+    , "replacement" .= repl
+    ] ++ maybe [] (\f -> ["flags" .= f]) flags
+
+instance FromJSON CharFilterDefinition where
+  parseJSON = withObject "CharFilterDefinition" $ \m -> do
+    t <- m .: "type"
+    case (t :: Text) of
+      "mapping" -> CharFilterDefinitionMapping . M.fromList <$> ms
+        where
+          ms = m .: "mappings" >>= mapM parseMapping
+          parseMapping kv = case T.splitOn "=>" kv of
+            (k:vs) -> pure (T.strip k, T.strip $ T.concat vs)
+            _ -> fail "mapping is not of the format key => value"
+      "pattern_replace" -> CharFilterDefinitionPatternReplace
+        <$> m .: "pattern" <*> m .: "replacement" <*> m .:? "flags"
+      _ -> fail ("unrecognized character filter type: " ++ T.unpack t)
+
 -- | Token filters are used to create custom analyzers.
 data TokenFilterDefinition
   = TokenFilterDefinitionLowercase (Maybe Language)
@@ -646,6 +690,8 @@ data TokenFilterDefinition
   | TokenFilterDefinitionReverse
   | TokenFilterDefinitionSnowball Language
   | TokenFilterDefinitionShingle Shingle
+  | TokenFilterDefinitionStemmer Language
+  | TokenFilterDefinitionStop (Either Language [StopWord])
   deriving (Eq,Show,Generic)
 
 instance ToJSON TokenFilterDefinition where
@@ -677,6 +723,16 @@ instance ToJSON TokenFilterDefinition where
       , "token_separator" .= shingleTokenSeparator s
       , "filler_token" .= shingleFillerToken s
       ]
+    TokenFilterDefinitionStemmer lang -> object
+      [ "type" .= ("stemmer" :: Text)
+      , "language" .= languageToText lang
+      ]
+    TokenFilterDefinitionStop stop -> object
+      [ "type" .= ("stop" :: Text)
+      , "stopwords" .= case stop of
+          Left lang -> String $ "_" <> languageToText lang <> "_"
+          Right stops -> toJSON stops
+      ]
 
 instance FromJSON TokenFilterDefinition where
   parseJSON = withObject "TokenFilterDefinition" $ \m -> do
@@ -697,6 +753,14 @@ instance FromJSON TokenFilterDefinition where
         <*> (fmap.fmap) unStringlyTypedBool (m .:? "output_unigrams_if_no_shingles") .!= False
         <*> m .:? "token_separator" .!= " "
         <*> m .:? "filler_token" .!= "_"
+      "stemmer" -> TokenFilterDefinitionStemmer
+        <$> m .: "language"
+      "stop" -> do
+        stop <- m .: "stopwords"
+        stop' <- case stop of
+          String lang -> fmap Left . parseJSON . String . T.drop 1 . T.dropEnd 1 $ lang
+          _ -> Right <$> parseJSON stop
+        return (TokenFilterDefinitionStop stop')
       _ -> fail ("unrecognized token filter type: " ++ T.unpack t)
 
 -- | The set of languages that can be passed to various analyzers,
@@ -1353,6 +1417,8 @@ newtype Tokenizer =
   Tokenizer Text deriving (Eq, Read, Show, Generic, ToJSON, FromJSON, Typeable)
 newtype TokenFilter =
   TokenFilter Text deriving (Eq, Read, Show, Generic, ToJSON, FromJSON, Typeable)
+newtype CharFilter =
+  CharFilter Text deriving (Eq, Read, Show, Generic, ToJSON, FromJSON, Typeable)
 newtype MaxExpansions =
   MaxExpansions Int deriving (Eq, Read, Show, Generic, ToJSON, FromJSON, Typeable)
 
@@ -1394,7 +1460,7 @@ newtype TypeName =
 newtype PercentMatch =
   PercentMatch Double deriving (Eq, Read, Show, Generic, ToJSON, FromJSON, Typeable)
 newtype StopWord =
-  StopWord Text deriving (Eq, Read, Show, Generic, ToJSON, FromJSON, Typeable)
+  StopWord Text deriving (Eq, Read, Show, Generic, ToJSON, FromJSON, Typeable, IsString)
 newtype QueryPath =
   QueryPath Text deriving (Eq, Read, Show, Generic, ToJSON, FromJSON, Typeable)
 
@@ -5711,4 +5777,4 @@ instance FromJSON DirectGenerators where
                       <*> o .:? "post_filter"
 
 mkDirectGenerators :: FieldName -> DirectGenerators
-mkDirectGenerators fn = DirectGenerators fn Nothing DirectGeneratorSuggestModeMissing Nothing Nothing Nothing Nothing Nothing Nothing Nothing Nothing 
+mkDirectGenerators fn = DirectGenerators fn Nothing DirectGeneratorSuggestModeMissing Nothing Nothing Nothing Nothing Nothing Nothing Nothing Nothing