+++ changelog.md
@@ -1,3 +1,12 @@
+UNRELEASED
+========
+- @rvl
+  - Add [character filters][] to the `AnalyzerDefinition` type.
+  - Add the "stemmer" and "stop" [token filters][] to `TokenFilterDefinition`.
+
+[Character Filters]: https://www.elastic.co/guide/en/elasticsearch/reference/5.6/analysis-charfilters.html
+[Token Filters]: https://www.elastic.co/guide/en/elasticsearch/reference/5.6/analysis-tokenfilters.html
+
 0.15.0.2
 ========
 - @michaelxavier